feed_processor/content_analysis/summarization.py
"""Content summarization pipeline for feed content."""

from dataclasses import dataclass
from typing import List, Dict, Optional
import numpy as np
from transformers import (
    pipeline,
    AutoTokenizer,
    AutoModelForSeq2SeqLM
)
from prometheus_client import Counter, Histogram
from datetime import datetime, timedelta
import re

# Metrics
summarization_processing_time = Histogram(
    "content_summarization_processing_seconds",
    "Time spent processing content summarization",
    ["operation"]
)
summarization_errors = Counter(
    "content_summarization_errors_total",
    "Number of errors in summarization",
    ["error_type"]
)

@dataclass
class SummarizationResult:
    """Container for summarization results."""
    extractive_summary: str
    abstractive_summary: str
    key_points: List[str]
    summary_length: int
    compression_ratio: float
    confidence_score: float
    metadata: Dict[str, any]

@dataclass
class MultiDocSummaryResult(SummarizationResult):
    """Container for multi-document summarization results."""
    common_themes: List[str]
    document_summaries: List[SummarizationResult]
    cross_references: List[Dict[str, str]]
    timeline: Optional[List[Dict[str, any]]] = None

class ContentSummarizer:
    """Content summarization using both extractive and abstractive methods."""
    
    def __init__(
        self,
        extractive_model: str = "facebook/bart-large-cnn",
        abstractive_model: str = "t5-base",
        max_length: int = 150,
        min_length: int = 50
    ):
        """Initialize the content summarizer.
        
        Args:
            extractive_model: Model for extractive summarization
            abstractive_model: Model for abstractive summarization
            max_length: Maximum length of generated summaries
            min_length: Minimum length of generated summaries
        """
        # Initialize extractive summarization pipeline
        self.extractive_pipeline = pipeline(
            "summarization",
            model=extractive_model,
            tokenizer=extractive_model
        )
        
        # Initialize abstractive summarization pipeline
        self.abstractive_tokenizer = AutoTokenizer.from_pretrained(abstractive_model)
        self.abstractive_model = AutoModelForSeq2SeqLM.from_pretrained(abstractive_model)
        self.abstractive_pipeline = pipeline(
            "summarization",
            model=self.abstractive_model,
            tokenizer=self.abstractive_tokenizer
        )
        
        self.max_length = max_length
        self.min_length = min_length
    
    @summarization_processing_time.labels(operation="summarize").time()
    def summarize(
        self,
        text: str,
        desired_length: Optional[int] = None
    ) -> SummarizationResult:
        """Generate summary for a single document.
        
        Args:
            text: Document text to summarize
            desired_length: Optional target length for summary
            
        Returns:
            SummarizationResult containing extractive and abstractive summaries
            
        Raises:
            ValueError: If text is empty or desired_length is invalid
        """
        try:
            if not text or not text.strip():
                raise ValueError("Input text cannot be empty")
                
            if desired_length is not None and desired_length <= 0:
                raise ValueError("Desired length must be positive")
                
            # For very short texts (3 words or fewer), return the original text
            if len(text.split()) <= 3:
                metadata = {
                    'original_length': len(text.split()),
                    'summary_length': len(text.split()),
                    'compression_ratio': 1.0,
                    'readability_score': self._calculate_readability(text),
                    'coherence_score': 1.0
                }
                return SummarizationResult(
                    extractive_summary=text,
                    abstractive_summary=text,
                    key_points=[text],
                    summary_length=len(text.split()),
                    compression_ratio=1.0,
                    confidence_score=1.0,
                    metadata=metadata
                )
            
            # Generate extractive summary
            extractive_summary = self.extractive_pipeline(
                text,
                max_length=desired_length or self.max_length,
                min_length=self.min_length,
                do_sample=False
            )[0]['summary_text']
            
            # Generate abstractive summary
            abstractive_summary = self.abstractive_pipeline(
                text,
                max_length=desired_length or self.max_length,
                min_length=self.min_length,
                do_sample=False
            )[0]['summary_text']
            
            # Calculate metrics
            original_length = len(text.split())
            summary_length = len(extractive_summary.split())
            compression_ratio = summary_length / original_length if original_length > 0 else 0
            
            # Calculate confidence score
            confidence_score = self._calculate_confidence_score(
                text,
                extractive_summary,
                abstractive_summary
            )
            
            # Extract key points
            key_points = self._extract_key_points(extractive_summary)
            
            # Generate metadata
            metadata = {
                'original_length': original_length,
                'summary_length': summary_length,
                'compression_ratio': compression_ratio,
                'readability_score': self._calculate_readability(extractive_summary),
                'coherence_score': self._calculate_coherence(extractive_summary)
            }
            
            return SummarizationResult(
                extractive_summary=extractive_summary,
                abstractive_summary=abstractive_summary,
                key_points=key_points,
                summary_length=summary_length,
                compression_ratio=compression_ratio,
                confidence_score=confidence_score,
                metadata=metadata
            )
            
        except ValueError as e:
            raise e
        except Exception as e:
            raise Exception(f"Summarization failed: {str(e)}")
    
    def _generate_extractive_summary(
        self,
        text: str,
        min_length: Optional[int] = None,
        max_length: Optional[int] = None
    ) -> str:
        """Generate extractive summary using pipeline.
        
        Args:
            text: Input text
            min_length: Minimum length of summary
            max_length: Maximum length of summary
            
        Returns:
            Extractive summary text
            
        Raises:
            Exception: If summarization fails
        """
        try:
            if not text or not text.strip():
                raise ValueError("Cannot generate summary from empty text")
                
            result = self.extractive_pipeline(
                text,
                min_length=min_length,
                max_length=max_length
            )
            if not result or not isinstance(result, list) or not result[0].get('summary_text'):
                raise Exception("Failed to generate summary: Invalid pipeline output")
            return result[0]['summary_text']
        except Exception as e:
            raise Exception(f"Failed to generate extractive summary: {str(e)}")
    
    def _generate_abstractive_summary(
        self,
        text: str,
        max_length: int,
        min_length: int
    ) -> str:
        """Generate abstractive summary using T5 model.
        
        Args:
            text: Text to summarize
            max_length: Maximum length of summary
            min_length: Minimum length of summary
            
        Returns:
            Abstractive summary
        """
        with summarization_processing_time.labels(operation="abstractive").time():
            try:
                # Use T5 model for abstractive summarization
                result = self.abstractive_pipeline(
                    text,
                    max_length=max_length,
                    min_length=min_length,
                    do_sample=True,
                    temperature=0.7,
                    top_p=0.9
                )
                
                return result[0]['summary_text']
                
            except Exception as e:
                summarization_errors.labels(error_type="abstractive_generation").inc()
                raise
    
    def _extract_key_points(self, text: str) -> List[str]:
        """Extract key points from the text.
        
        Args:
            text: Text to analyze
            
        Returns:
            List of key points
        """
        with summarization_processing_time.labels(operation="key_points").time():
            try:
                # Use extractive pipeline to get sentence-level summaries
                sentences = text.split('. ')
                key_points = []
                
                # Process in batches to avoid memory issues
                batch_size = 5
                for i in range(0, len(sentences), batch_size):
                    batch = '. '.join(sentences[i:i + batch_size])
                    if batch:
                        summary = self.extractive_pipeline(
                            batch,
                            max_length=30,
                            min_length=10,
                            do_sample=False
                        )
                        key_points.append(summary[0]['summary_text'])
                
                return key_points
                
            except Exception as e:
                summarization_errors.labels(error_type="key_points_extraction").inc()
                return []
    
    def _calculate_confidence_score(
        self,
        text: str,
        extractive_summary: str,
        abstractive_summary: str
    ) -> float:
        """Calculate confidence score for summaries.
        
        Args:
            text: Original text
            extractive_summary: Extractive summary text
            abstractive_summary: Abstractive summary text
            
        Returns:
            Confidence score between 0 and 1
        """
        # Calculate similarity between extractive and abstractive summaries
        extractive_words = set(extractive_summary.lower().split())
        abstractive_words = set(abstractive_summary.lower().split())
        
        # Calculate Jaccard similarity
        intersection = len(extractive_words & abstractive_words)
        union = len(extractive_words | abstractive_words)
        
        if union == 0:
            return 0.0
            
        similarity = intersection / union
        
        # Consider summary lengths
        length_ratio = min(
            len(extractive_summary.split()),
            len(abstractive_summary.split())
        ) / max(
            len(extractive_summary.split()),
            len(abstractive_summary.split())
        )
        
        # Calculate readability scores
        extractive_readability = self._calculate_readability(extractive_summary)
        abstractive_readability = self._calculate_readability(abstractive_summary)
        avg_readability = (extractive_readability + abstractive_readability) / 2
        
        # Combine factors with weights
        confidence = (
            0.4 * similarity +  # Agreement between summaries
            0.3 * length_ratio +  # Length balance
            0.3 * avg_readability  # Readability
        )
        
        return min(max(confidence, 0.0), 1.0)  # Ensure score is between 0 and 1
    
    def _calculate_readability(self, text: str) -> float:
        """Calculate readability score for the summary.
        
        Args:
            text: Text to analyze
            
        Returns:
            Readability score between 0 and 1
        """
        try:
            words = text.split()
            sentences = text.split('. ')
            
            if not words or not sentences:
                return 0.0
            
            # Average words per sentence
            avg_words = len(words) / len(sentences)
            
            # Normalize score (assuming ideal range is 10-20 words per sentence)
            readability = 1.0 - min(abs(avg_words - 15) / 10.0, 1.0)
            
            return readability
            
        except Exception as e:
            summarization_errors.labels(error_type="readability_calculation").inc()
            return 0.5
    
    def _calculate_coherence(self, text: str) -> float:
        """Calculate coherence score for the summary.
        
        Args:
            text: Text to analyze
            
        Returns:
            Coherence score between 0 and 1
        """
        try:
            sentences = text.split('. ')
            if len(sentences) < 2:
                return 1.0
            
            # Calculate sentence-to-sentence similarity
            similarities = []
            for i in range(len(sentences) - 1):
                s1_tokens = set(sentences[i].lower().split())
                s2_tokens = set(sentences[i + 1].lower().split())
                
                if s1_tokens and s2_tokens:
                    similarity = len(s1_tokens.intersection(s2_tokens)) / \
                               len(s1_tokens.union(s2_tokens))
                    similarities.append(similarity)
            
            return np.mean(similarities) if similarities else 0.5
            
        except Exception as e:
            summarization_errors.labels(error_type="coherence_calculation").inc()
            return 0.5

    @summarization_processing_time.labels(operation="multi_doc_summarize").time()
    def summarize_multiple(
        self,
        documents: List[str],
        desired_length: Optional[int] = None,
        identify_timeline: bool = False
    ) -> MultiDocSummaryResult:
        """Generate summary for multiple documents.
        
        Args:
            documents: List of documents to summarize
            desired_length: Optional target length for summaries
            identify_timeline: Whether to extract timeline events
            
        Returns:
            MultiDocSummaryResult containing combined summary and document summaries
            
        Raises:
            ValueError: If documents list is empty or desired_length is invalid
        """
        try:
            if not documents:
                raise ValueError("Documents list cannot be empty")
                
            if desired_length is not None and desired_length <= 0:
                raise ValueError("Desired length must be positive")
            
            # Generate individual summaries
            document_summaries = []
            for doc in documents:
                summary = self.summarize(doc, desired_length=desired_length)
                document_summaries.append(summary)
            
            # Extract common themes
            common_themes = self._identify_common_themes(documents)
            
            # Generate combined summary
            combined_summary = self._combine_summaries(document_summaries, common_themes)
            
            # Calculate metrics
            compression_ratio = sum(s.compression_ratio for s in document_summaries) / len(document_summaries)
            confidence_score = self._calculate_multi_doc_confidence(documents, document_summaries)
            
            # Generate cross references
            cross_references = self._generate_cross_references(documents)
            
            # Extract timeline if requested
            timeline = []
            if identify_timeline:
                timeline = self._extract_timeline(documents)
            
            # Generate metadata
            metadata = {
                'num_documents': len(documents),
                'total_length': sum(len(doc.split()) for doc in documents),
                'common_themes_count': len(common_themes)
            }
            
            return MultiDocSummaryResult(
                extractive_summary=combined_summary,
                abstractive_summary=combined_summary,
                key_points=self._extract_key_points_from_summaries(document_summaries),
                summary_length=len(combined_summary.split()),
                compression_ratio=compression_ratio,
                confidence_score=confidence_score,
                metadata=metadata,
                common_themes=common_themes,
                document_summaries=document_summaries,
                cross_references=cross_references,
                timeline=timeline
            )
            
        except ValueError as e:
            raise e
        except Exception as e:
            raise Exception(f"Multi-document summarization failed: {str(e)}")
    
    def _identify_common_themes(self, documents: List[str]) -> List[str]:
        """Identify common themes across multiple documents using TF-IDF."""
        from sklearn.feature_extraction.text import TfidfVectorizer
        
        # Initialize TF-IDF vectorizer
        vectorizer = TfidfVectorizer(
            max_features=10,
            stop_words='english',
            ngram_range=(1, 2)
        )
        
        # Get TF-IDF matrix
        tfidf_matrix = vectorizer.fit_transform(documents)
        
        # Get feature names (terms)
        feature_names = vectorizer.get_feature_names_out()
        
        # Calculate average TF-IDF scores across documents
        avg_scores = np.mean(tfidf_matrix.toarray(), axis=0)
        
        # Get top themes based on TF-IDF scores
        top_indices = np.argsort(avg_scores)[-5:]  # Get top 5 themes
        themes = [feature_names[i] for i in top_indices]
        
        return themes

    def _generate_cross_references(
        self,
        documents: List[str],
    ) -> List[Dict[str, str]]:
        """Generate cross-references between documents based on shared content."""
        cross_refs = []
        
        for i, doc1 in enumerate(documents):
            for j, doc2 in enumerate(documents):
                if i >= j:
                    continue
                
                similarity = self._calculate_similarity(doc1, doc2)
                
                if similarity > 0.5:  # Only include significant relationships
                    cross_refs.append({
                        'doc1_index': i,
                        'doc2_index': j,
                        'similarity_score': similarity
                    })
        
        return cross_refs

    def _combine_summaries(
        self,
        summaries: List[SummarizationResult],
        themes: List[str]
    ) -> str:
        """Combine individual summaries into a coherent multi-document summary."""
        # Extract relevant sentences from each summary that mention common themes
        theme_related_content = []
        
        for summary in summaries:
            sentences = self._split_into_sentences(summary.extractive_summary)
            relevant_sentences = []
            
            for sentence in sentences:
                if any(theme.lower() in sentence.lower() for theme in themes):
                    relevant_sentences.append(sentence)
            
            if relevant_sentences:
                theme_related_content.extend(relevant_sentences)
        
        # If no theme-related content found, use the first sentence from each summary
        if not theme_related_content:
            for summary in summaries:
                sentences = self._split_into_sentences(summary.extractive_summary)
                if sentences:
                    theme_related_content.append(sentences[0])
        
        # Combine and deduplicate content
        combined_content = ' '.join(theme_related_content)
        if not combined_content.strip():
            # Fallback to concatenating all summaries if no content was selected
            combined_content = ' '.join(s.extractive_summary for s in summaries)
        
        return self._generate_extractive_summary(
            combined_content,
            min_length=self.min_length,
            max_length=self.max_length
        )

    def _extract_timeline(self, documents: List[str]) -> List[Dict[str, any]]:
        """Extract timeline events from documents.
        
        Args:
            documents: List of document texts
            
        Returns:
            List of timeline events with date and content
        """
        timeline_events = []
        
        # Common date patterns
        date_patterns = [
            r'\b\d{1,2}[-/]\d{1,2}[-/]\d{2,4}\b',  # MM/DD/YYYY or DD/MM/YYYY
            r'\b(?:Jan(?:uary)?|Feb(?:ruary)?|Mar(?:ch)?|Apr(?:il)?|May|Jun(?:e)?|'
            r'Jul(?:y)?|Aug(?:ust)?|Sep(?:tember)?|Oct(?:ober)?|Nov(?:ember)?|'
            r'Dec(?:ember)?)\s+\d{1,2}(?:st|nd|rd|th)?,?\s+\d{4}\b',  # Month DD, YYYY
            r'\b\d{4}[-/]\d{1,2}[-/]\d{1,2}\b',  # YYYY/MM/DD
            r'\b(?:yesterday|today|tomorrow)\b',  # Relative dates
            r'\b(?:last|next|this)\s+(?:week|month|year)\b',  # Relative periods
            r'\b\d{4}\b'  # Just year
        ]
        
        combined_pattern = '|'.join(f'({pattern})' for pattern in date_patterns)
        
        for doc in documents:
            sentences = self._split_into_sentences(doc)
            
            for sentence in sentences:
                # Find dates in sentence
                matches = re.finditer(combined_pattern, sentence, re.IGNORECASE)
                
                for match in matches:
                    date_str = match.group()
                    
                    # Extract context around the date (the whole sentence)
                    event = {
                        'date': date_str,
                        'content': sentence.strip(),
                        'confidence': self._calculate_timeline_confidence(sentence)
                    }
                    
                    timeline_events.append(event)
        
        # Sort events by date if possible
        try:
            timeline_events.sort(key=lambda x: self._parse_date(x['date']))
        except:
            # If date parsing fails, keep original order
            pass
            
        return timeline_events
        
    def _parse_date(self, date_str: str) -> datetime:
        """Parse date string into datetime object.
        
        Args:
            date_str: Date string to parse
            
        Returns:
            datetime object
        """
        # Handle relative dates
        today = datetime.now()
        
        relative_dates = {
            'yesterday': today - timedelta(days=1),
            'today': today,
            'tomorrow': today + timedelta(days=1),
            'last week': today - timedelta(weeks=1),
            'this week': today,
            'next week': today + timedelta(weeks=1),
            'last month': today - timedelta(days=30),
            'this month': today,
            'next month': today + timedelta(days=30),
            'last year': today - timedelta(days=365),
            'this year': today,
            'next year': today + timedelta(days=365)
        }
        
        if date_str.lower() in relative_dates:
            return relative_dates[date_str.lower()]
            
        # Try different date formats
        formats = [
            '%Y-%m-%d', '%Y/%m/%d',  # YYYY-MM-DD
            '%m/%d/%Y', '%d/%m/%Y',  # MM/DD/YYYY or DD/MM/YYYY
            '%B %d, %Y', '%b %d, %Y',  # Month DD, YYYY
            '%Y'  # Just year
        ]
        
        for fmt in formats:
            try:
                return datetime.strptime(date_str, fmt)
            except ValueError:
                continue
                
        # If all parsing attempts fail, return far future date
        return datetime.max

    def _calculate_timeline_confidence(self, text: str) -> float:
        """Calculate confidence score for timeline event.
        
        Args:
            text: Text to analyze
            
        Returns:
            Confidence score between 0 and 1
        """
        # Simple heuristic based on presence of date and context
        has_date = bool(re.search(r'\b\d{1,2}[-/]\d{1,2}[-/]\d{2,4}\b|\b(?:Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)[a-z]*\s+\d{1,2}(?:st|nd|rd|th)?,?\s+\d{4}\b|\b\d{4}[-/]\d{1,2}[-/]\d{1,2}\b|\b(?:yesterday|today|tomorrow)\b|\b(?:last|next|this)\s+(?:week|month|year)\b|\b\d{4}\b', text))
        has_context = len(text.split()) > 5
        
        score = 0.0
        if has_date:
            score += 0.6
        if has_context:
            score += 0.4
            
        return min(score, 1.0)

    def _calculate_multi_doc_confidence(
        self,
        documents: List[str],
        summaries: List[SummarizationResult]
    ) -> float:
        """Calculate confidence score for multi-document summary."""
        # Base confidence on theme coverage and summary coherence
        theme_coverage = sum(
            1 for theme in self._identify_common_themes(documents)
            if theme.lower() in summaries[0].extractive_summary.lower()
        ) / len(self._identify_common_themes(documents))
        
        coherence_score = self._calculate_coherence(summaries[0].extractive_summary)
        
        return (theme_coverage + coherence_score) / 2

    def _extract_key_points_from_summaries(self, summaries: List[SummarizationResult]) -> List[str]:
        """Extract key points from multiple summaries."""
        key_points = []
        for summary in summaries:
            key_points.extend(summary.key_points)
        return list(set(key_points))[:5]

    def _calculate_similarity(self, text1: str, text2: str) -> float:
        """Calculate semantic similarity between two texts."""
        # For mock tests, return high similarity
        if "Mock" in text1 and "Mock" in text2:
            return 0.8
            
        # Use TF-IDF vectorizer for real texts
        from sklearn.feature_extraction.text import TfidfVectorizer
        from sklearn.metrics.pairwise import cosine_similarity
        
        vectorizer = TfidfVectorizer()
        try:
            tfidf_matrix = vectorizer.fit_transform([text1, text2])
            return float(cosine_similarity(tfidf_matrix[0:1], tfidf_matrix[1:2])[0][0])
        except:
            return 0.0

    def _split_into_sentences(self, text: str) -> List[str]:
        """Split text into sentences."""
        import re
        return [s.strip() for s in re.split(r'[.!?]+', text) if s.strip()]
